refactor(renderer): MetaGroup rAF 扫光引擎抽 use-sweep-highlight，summaryLabel 挪 meta-summary-label
